implemented PSR-1 & PSR-2 formatting, refactored TicTacToe, removed any output in functions

+++ 01-Basics/01-Arrays/03-Exercise.php
@@ -7,9 +7,8 @@ $numbers = [
 ];
 
 echo "Enter the value to search for: ";
-$value = readline();
+$value = (int) readline();
 
-//todo check if an array contains a value user entered
 if (in_array($value, $numbers)) {
     echo "The value $value is in the array.\n";
 } else {
