+++ 01-Basics/04-Flow/02-Exercise.php
@@ -1,10 +1,8 @@
 <?php
 
-
 echo "Enter the number: ";
-$number = readline();
+$number = (int) readline();
 
-//todo print if number is positive or negative
 if ($number > 0) {
     echo "Number $number is positive.\n";
 } elseif ($number < 0) {
